+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Dependency-Track Jenkins Plugin - Changelog
 
 ## Unreleased
+### ⭐ New Features
+- allow to specify an alternative URL for the Dependency-Track Frontend (fixes [#22](https://github.com/jenkinsci/dependency-track-plugin/issues/22))
+
 ### 🐞 Bugs Fixed
 - verify that `projectId` is set when auto create projects is not enabled